Rename models.py -> meta.py
